requirements.txt
spade==4.1.0
Flask==2.2.5
jinja2==3.0.3
openai==1.30.1
python-dotenv==0.28y